JAMES-1854 Extract poster
