+++ server/mailet/mailets/src/test/java/org/apache/james/transport/mailets/delivery/SieveMailetTest.java
@@ -107,6 +107,7 @@ public class SieveMailetTest {
             .userRepository(usersRepository)
             .mailboxManager(mailboxManager)
             .resourceLocator(resourceLocator)
+            .mailetContext(fakeMailContext)
             .folder("INBOX")
             .consume(true)
             .build();
